Move ecliptic definitions to new module frames
